fix: PostPage.tsxでtemplate化する
